Fix bug in SeismicReader.gen_cdp_y
the cdp_y coords were being generated off of the x coord of the origin rather than the y.

This was not noticeable in the test data, so I added a translated and rotated version of the 32-bit test data. Prior to this fix, the tests were failing on the new volumes